refactor: FarmPersonaSelect 분리
